Fase 3: bloques de 16 bits, formas, colocación, puertas y camas
- Ids de bloque de 16 bits (hasta 4096 estados) con migración automática de los
  mundos guardados (Durable Object e IndexedDB); protocolo v4.
- Familias de bloques con propiedades y modelos de cajas: losas (y dobles) y
  escaleras de 8 materiales, vallas, portillos, puertas con bisagra, trampillas,
  escaleras de mano, paneles de cristal, antorchas en la pared y cama.
- Colisiones por cajas con subida de escalones de 0,6 (jugador y criaturas),
  trepar escaleras de mano, raycast y grietas con la forma real.
- Colocación compartida cliente/servidor (autoritativa): mitades, orientación,
  bisagra, puertas dobles; abrir y cerrar; apoyos de vecinos y botín único.
- Camas: dormir de noche con todos los jugadores (no con monstruos cerca),
  amanecer y reaparición en la cama.
- Recetas, iconos y modelos 3D en la mano, texturas de puertas, trampillas y
  escalera de mano, sprites de puertas y cama.
- Pruebas en el repositorio (npm test con node:test) y GitHub Actions.

+++ src/client/world/World.ts
@@ -13,7 +13,7 @@ export interface Column {
   cx: number;
   cz: number;
   key: string;
-  blocks: Uint8Array | null;
+  blocks: Uint16Array | null;
   light: Uint8Array | null;
   /** y máximo con bloques (para truncar las copias hacia los workers). */
   maxY: number;
